// src/api/estrutura/models/ResEstruturaItemTreeDTO.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { ResEstruturaItemDTO } from './ResEstruturaItemDTO';

export type ResEstruturaItemTreeDTO = {
  /**
   * @type string | undefined
   */
  status?: string;
  /**
   * @type string | undefined
   */
  itemCliente?: string;
  /**
   * @type string | undefined
   */
  partcode?: string;
  /**
   * @type number | undefined
   */
  qtd?: number;
  /**
   * @type boolean | undefined
   */
  ehControle?: boolean;
  /**
   * @type array | undefined
   */
  children?: ResEstruturaItemTreeDTO[];
  /**
   * @type object | undefined
   */
  father?: ResEstruturaItemDTO;
};